+++ web/src/client/css/layout/_forms.scss
@@ -85,7 +85,7 @@ form {
   input[type='text'],
   input[type='email'],
   input[type='number'],
-  .StripeElement .InputContainer {
+  .StripeElement {
     background: var(--form-background);
     border: thin solid var(--base-border-color);
     border-radius: $borderRadius;
