+++ web/src/client/js/components/AdminDashboard/AdminBilling/StripeContainer.js
@@ -1,8 +1,12 @@
 import React from 'react'
+import { Redirect } from 'react-router-dom'
 import { StripeProvider } from 'react-stripe-elements'
-import BillingComponent from './BillingComponent'
 
-class BillingContainer extends React.Component {
+import { ORGANIZATION_ROLE_IDS, PATH_PROJECTS } from 'Shared/constants'
+import OrgPermission from 'Components/Permission/OrgPermission'
+import StripeComponent from './StripeComponent'
+
+class StripeContainer extends React.Component {
   constructor(props) {
     super(props)
     this.state = {
@@ -23,11 +27,11 @@ class BillingContainer extends React.Component {
 
   render() {
     return (
-      <div className="billing-container">
+      <OrgPermission acceptedRoleIds={[ORGANIZATION_ROLE_IDS.OWNER]} elseRender={<Redirect to={PATH_PROJECTS} />}>
         <StripeProvider stripe={this.state.stripe}>
-          <BillingComponent />
+          <StripeComponent />
         </StripeProvider>
-      </div>
+      </OrgPermission>
     )
   }
 
@@ -47,4 +51,4 @@ class BillingContainer extends React.Component {
   }
 }
 
-export default BillingContainer
+export default StripeContainer
