+++ web/src/client/js/components/AdminDashboard/AdminPayment/AdminPaymentContainer.js
@@ -1,0 +1,12 @@
+import React from 'react'
+
+import AdminPaymentComponent from './AdminPaymentComponent'
+
+const AdminPaymentContainer = () => {
+  return <AdminPaymentComponent />
+}
+
+AdminPaymentContainer.propTypes = {
+}
+
+export default AdminPaymentContainer
