+++ web/src/client/js/components/AdminDashboard/AdminSeats/AdminSeatsContainer.js
@@ -1,0 +1,12 @@
+import React from 'react'
+
+import AdminSeatsComponent from './AdminSeatsComponent'
+
+const AdminSeatsContainer = () => {
+  return <AdminSeatsComponent />
+}
+
+AdminSeatsContainer.propTypes = {
+}
+
+export default AdminSeatsContainer
