+++ web/src/client/js/components/AdminDashboard/AdminSeats/AdminSeatsStyles.scss
@@ -1,0 +1,36 @@
+@import 'CSS/inc/_vars.scss';
+
+.admin-seats {
+  &__status {
+    &__label {
+      display: block;
+      font-size: 1.2rem;
+    }
+    &__meter {
+      // -webkit-appearance: none; // fixes chrome / opera
+      -moz-appearance: none;
+      background: $gray-light; // Firefox
+      display: block;
+      margin: .25rem 0 1rem 0;
+      width: 100%;
+      &::-webkit-meter-bar {
+        background: $gray-light;
+        border-radius: $borderRadius;
+      }
+      &::-webkit-meter-optimum-value,
+      &::-webkit-meter-suboptimum-value,
+      &::-webkit-meter-even-less-good-value {
+        height: 100%;
+      }
+      &::-webkit-meter-optimum-value {
+        background: $primary;
+      }
+      &::-webkit-meter-suboptimum-value {
+        background: $highlight;
+      }
+      &::-webkit-meter-even-less-good-value {
+        background: $danger;
+      }
+    }
+  }
+}
