--- web/src/client/js/components/Billing/BillingStyles.scss
@@ -1,1 +1,0 @@
-// Import global billing styles, used outside of JS as well
